+++ sample/structure/lib/app/module/main/controller/main_controller.dart
@@ -1,0 +1,11 @@
+import '../../../utils/import.dart';
+
+class MainController extends GetxController {
+  RxInt selectedIndex = 0.obs;
+  RxBool isLoading = true.obs;
+
+  void changeTab(int index) {
+    if (selectedIndex.value == index) return;
+    selectedIndex.value = index;
+  }
+}