+++ sample/structure/lib/app/routes/app_page.dart
@@ -1,0 +1,21 @@
+import '../module/main/main_export.dart';
+import '../module/splash/splash_export.dart';
+import '../utils/import.dart';
+
+class AppPage {
+  AppPage._();
+
+  static final routers = <GetPage>[
+    GetPage(
+      name: AppRoute.splashView,
+      page: () => const SplashView(),
+      binding: SplashBinding(),
+    ),
+
+    GetPage(
+      name: AppRoute.mainView,
+      page: () => const MainView(),
+      binding: MainBinding(),
+    ),
+  ];
+}