+++ sample/structure/lib/app/routes/app_route.dart
@@ -1,0 +1,4 @@
+abstract class AppRoute {
+  static const splashView = '/SplashView';
+  static const mainView = '/MainView';
+}