+++ sample/structure/lib/app/utils/image_assets.dart
@@ -1,0 +1,16 @@
+class ImageAssetPNG {
+  const ImageAssetPNG._();
+}
+
+class ImageAssetSVG {
+  const ImageAssetSVG._();
+}
+
+class ImageAssetJPG {
+  const ImageAssetJPG._();
+}
+
+class IconAssetSVG {
+  const IconAssetSVG._();
+  static const String back = 'assets/icons/svg/back.svg';
+}